ofxMultiTouch now takes float instead of int
git-svn-id: http://ofxmsaof.googlecode.com/svn/trunk@81 a9ddf6a8-909b-11dd-94ae-b329dc30fcad

+++ ofxMultiTouch/src/ofxMultiTouch.h
@@ -36,7 +36,7 @@ public:
 	
 	// multitouch engine should call this when a 'down' is detected
 	// use ofxMultiTouchCustomData *data to send any platform specific data
-	void touchDown(int x, int y, int touchId, ofxMultiTouchCustomData *data = 0) {
+	void touchDown(float x, float y, int touchId, ofxMultiTouchCustomData *data = 0) {
 		for(std::list<ofxMultiTouchListener*>::iterator it=listeners.begin(); it!=listeners.end(); ++it) {
 			ofxMultiTouchListener* o = *it;
 			o->touchDown(x, y, touchId, data);
@@ -45,7 +45,7 @@ public:
 	
 	// multitouch engine should call this when a 'move' (drag) is detected
 	// use ofxMultiTouchCustomData *data to send any platform specific data
-	void touchMoved(int x, int y, int touchId, ofxMultiTouchCustomData *data = 0) {
+	void touchMoved(float x, float y, int touchId, ofxMultiTouchCustomData *data = 0) {
 		for(std::list<ofxMultiTouchListener*>::iterator it=listeners.begin(); it!=listeners.end(); ++it) {
 			ofxMultiTouchListener* o = *it;
 			o->touchMoved(x, y, touchId, data);
@@ -54,7 +54,7 @@ public:
 	
 	// multitouch engine should call this when an 'up' is detected
 	// use ofxMultiTouchCustomData *data to send any platform specific data
-	void touchUp(int x, int y, int touchId, ofxMultiTouchCustomData *data = 0) {
+	void touchUp(float x, float y, int touchId, ofxMultiTouchCustomData *data = 0) {
 		for(std::list<ofxMultiTouchListener*>::iterator it=listeners.begin(); it!=listeners.end(); ++it) {
 			ofxMultiTouchListener* o = *it;
 			o->touchUp(x, y, touchId, data);
@@ -63,7 +63,7 @@ public:
 	
 	// multitouch engine should call this when a 'double tap' is detected
 	// use ofxMultiTouchCustomData *data to send any platform specific data	
-	void touchDoubleTap(int x, int y, int touchId, ofxMultiTouchCustomData *data = 0) {
+	void touchDoubleTap(float x, float y, int touchId, ofxMultiTouchCustomData *data = 0) {
 		for(std::list<ofxMultiTouchListener*>::iterator it=listeners.begin(); it!=listeners.end(); ++it) {
 			ofxMultiTouchListener* o = *it;
 			o->touchDoubleTap(x, y, touchId, data);
